// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
        Person: {
            readonly id: number;
            name: string;
        };
        PersonCreate: {
            name: string;
        };
        Expense: {
            readonly id: number;
            description: string;
            /** @description Currency amount, rounded to 2 decimal places. */
            amount: number;
            payer_id: number;
            /** Format: date */
            date: string;
            participant_ids: number[];
        };
        ExpenseWrite: {
            description: string;
            /** @description Currency amount, rounded to 2 decimal places. */
            amount: number;
            payer_id: number;
            /** Format: date */
            date: string;
            participant_ids: number[];
        };
        ExpensePatch: {
            description?: string;
            /** @description Currency amount, rounded to 2 decimal places. */
            amount?: number;
            payer_id?: number;
            /** Format: date */
            date?: string;
            participant_ids?: number[];
        };
        Balance: {
            person_id: number;
            name: string;
            balance: number;
        };
        Error: {
            detail: string;
        };
        HTTPValidationError: {
            detail: {
                loc: (string | number)[];
                msg: string;
                type: string;
            }[];
        };
    };
    responses: never;
    parameters: {
        ExpenseId: number;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}